requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# LangChain & LLM
langchain==0.1.0
langchain-community==0.0.10
langchain-openai==0.0.2
openai==1.8.0
tiktoken==0.5.2

# Embeddings & Vector Store
sentence-transformers==2.2.2
chromadb==0.4.22
hnswlib==0.8.0

# Document Processing
pypdf==3.17.4
python-docx==1.1.0
beautifulsoup4==4.12.2
markdown==3.5.1
unstructured==0.11.0

# Async & Background Tasks
celery==5.3.4
redis==5.0.1

# Observability
prometheus-client==0.19.0
structlog==23.3.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0